x86: kdump: move reserve_crashkernel[_low]() into crash_core.c
From: Chen Zhou <chenzhou10@huawei.com>
Link: https://lkml.org/lkml/2021/1/30/53
Link: https://github.com/openeuler-mirror/kernel/commit/8cb8686864a99cfa7962a56777535101d60f8e39

Make the functions reserve_crashkernel[_low]() as generic.
Arm64 will use these to reimplement crashkernel=X.

Tested-by: John Donnelly <John.p.donnelly@oracle.com>
Acked-by: Xie XiuQi <xiexiuqi@huawei.com>
